% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect_tissue_type.R
\name{auto_detect_tissue_type}
\alias{auto_detect_tissue_type}
\title{auto_detect_tissue_type}
\usage{
auto_detect_tissue_type(
  path_to_db_file,
  seuratObject,
  scaled,
  assay = "RNA",
  ...
)
}
\arguments{
\item{path_to_db_file}{DB file with cell types}

\item{seuratObject}{The Seurat Object from wich to extract the input scRNA-seq
matrix (rownames - genes, column names - cells)}

\item{scaled}{indicates whether the matrix is scaled (TRUE by default)}

\item{assay}{e.g. RNA, SCT, integrated}

\item{...}{TODO}
}
\value{
TODO
}
\description{
automatically detect a tissue type of the dataset
}
\examples{
# TODO
}
\author{
Aleksandr Ianevski, with edits from Federico Marini
}
