% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_neg.R
\name{get_pos_neg}
\alias{get_pos_neg}
\title{Extract postive and negative markers from DE results}
\usage{
get_pos_neg(ref_markers, cl, min_pct_diff)
}
\arguments{
\item{ref_markers}{dataframe with reference markers}

\item{cl}{celltype}

\item{min_pct_diff}{numeric}
}
\value{
list of positive and negative markers
}
\description{
Extract postive and negative markers from DE results
}
