% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bubbleplot.R
\name{make_bubbleplot}
\alias{make_bubbleplot}
\title{Make bubbleplot}
\usage{
make_bubbleplot(cL_results, sctype_scores, custom_marker_file)
}
\arguments{
\item{cL_results}{dataframe}

\item{sctype_scores}{scType scores}
}
\description{
Make bubbleplot
}
\author{
@IanevskiAleksandr
}
