% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_celltype_markers.R
\name{retrieve_markers}
\alias{retrieve_markers}
\title{Retrieve markers}
\usage{
retrieve_markers(
  ref_markers,
  tissue,
  marker_per_cluster = NULL,
  min_pct_diff = 0.3
)
}
\arguments{
\item{ref_markers}{dataframe with reference markers (results from DE)}

\item{tissue}{tissue type}

\item{marker_per_cluster}{no. marker genes per cluster to select}

\item{min_pct_diff}{min. percentage difference (default=0.3)}
}
\description{
Retrieve markers
}
