% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_score_.R
\name{sctype_score}
\alias{sctype_score}
\title{sctype_score}
\usage{
sctype_score(
  scRNAseqData,
  scaled = !0,
  gs,
  gs2 = NULL,
  gene_names_to_uppercase = !0,
  ...
)
}
\arguments{
\item{scRNAseqData}{input scRNA-seq matrix (rownames - genes, column names - cells)}

\item{scaled}{indicates whether the matrix is scaled (TRUE by default)}

\item{gs}{list of gene sets positively expressed in the cell type}

\item{gs2}{list of gene sets that should not be expressed in the cell type (NULL if not applicable)}

\item{gene_names_to_uppercase}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
calculate ScType scores and assign cell types
}
\examples{
# TODO
}
\author{
Aleksandr Ianevski, with edits from Federico Marini
}
