% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_wrapper.R
\name{sctype_source}
\alias{sctype_source}
\title{sctype source files}
\usage{
sctype_source()
}
\value{
original ScType database
}
\description{
loads sctype functions needed for an automated cell type annotation .
}
\details{
none
}
\examples{
db_ <- sctype_source()

}
\author{
Aleksandr Ianevski, with edits from Federico Marini
}
