% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect_tissue_type.R
\name{auto_detect_tissue_type}
\alias{auto_detect_tissue_type}
\title{Autotomatically detect a tissue of the dataset}
\usage{
auto_detect_tissue_type(
  celltype_markers_db,
  seuratObject,
  scaled,
  assay = "RNA",
  plot = FALSE,
  ...
)
}
\arguments{
\item{celltype_markers_db}{(dataframe) DB file with cell types}

\item{seuratObject}{Seurat Object from wich to extract the input scRNA-seq matrix (rownames - genes, column names - cells),}

\item{scaled}{logical indicating whether the matrix is scaled (default=TRUE)}

\item{assay}{e.g. RNA, SCT, integrated}

\item{plot}{logical indicating whether to plot (default=FALSE)}
}
\value{
list with two keys containing the postive and negative markers resp.
}
\description{
Autotomatically detect a tissue of the dataset
}
\author{
@IanevskiAleksandr
}
