% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets_prepare.R
\name{gene_sets_prepare}
\alias{gene_sets_prepare}
\title{Prepare gene sets}
\usage{
gene_sets_prepare(celltype_markers_db, cell_type)
}
\arguments{
\item{celltype_markers_db}{(dataframe) DB file with cell types}

\item{cell_type}{cell type (e.g. Immune system, Liver, Pancreas, Kidney, Eye, Brain)}
}
\value{
list with two keys containing the postive and negative markers resp.
}
\description{
Prepare gene sets
}
\author{
@IanevskiAleksandr
}
