% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_wrapper.R
\name{run_scType}
\alias{run_scType}
\alias{run_sctype}
\title{Run sctype analysis on Seurat object}
\usage{
run_sctype(
  seurat_object,
  known_tissue_type = NULL,
  assay = "RNA",
  scaled = TRUE,
  custom_marker_file = NULL,
  plot = FALSE,
  name = "sctype_classification"
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{known_tissue_type}{The tissue type of the input data (optional)}

\item{assay}{TODO}

\item{scaled}{TODO}

\item{custom_marker_file}{Path to the custom marker file (optional)}

\item{plot}{Whether to plot the results (default is FALSE)}

\item{name}{The name of the metadata column to store the scType results
(default is "sctype_classification")}
}
\value{
A modified copy of the input Seurat object with a new metadata column
}
\description{
run an automated cell type annotation
}
\details{
This function is compatible with seurat versions 4 and 5 --
different methods(@,$) to retrieve counts,data,scale.data
}
\examples{
# seurat_object <- run_scType(seurat_object, "Immune system")


}
\author{
Aleksandr Ianevski, with edits from Federico Marini
}
