% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sctype.R
\name{run_sctype}
\alias{run_sctype}
\title{Run sctype analysis on Seurat object}
\usage{
run_sctype(
  seurat_object,
  known_tissue_type = NULL,
  cluster_column = "seurat_clusters",
  assay = "RNA",
  scaled = TRUE,
  custom_marker_file = NULL,
  plot_umap = FALSE,
  plot_bubble = FALSE,
  name = "sctype_classification"
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{known_tissue_type}{The tissue type of the input data (optional)}

\item{cluster_column}{column in metadata containing the clusters (default='seurat_clusters')}

\item{assay}{e.g. RNA, SCT, integrated (default="RNA")}

\item{scaled}{logical indicating whether the matrix is scaled (default=TRUE)}

\item{custom_marker_file}{Path to the custom marker file (optional)}

\item{plot_umap}{logical indicating whether to plot the UMAP (default= FALSE)}

\item{plot_bubble}{logical indicating whether to plot the bubble plot (default=FALSE)}

\item{name}{The name of the metadata column to store the scType results (default="sctype_classification")}
}
\value{
A modified copy of the input Seurat object with a new metadata column
}
\description{
run an automated cell type annotation
}
\details{
This function is compatible with seurat versions 4 and 5 -- different methods(@,$) to retrieve counts,data,scale.data
}
\examples{
\dontrun{
seurat_object <- run_scType(seurat_object, "Immune system")
}
}
\author{
@IanevskiAleksandr
}
