% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptype.R
\name{run_scType}
\alias{run_scType}
\alias{run_sptype}
\title{Run sctype analysis on Seurat object}
\usage{
run_sptype(
  seurat_object,
  known_tissue_type = NULL,
  custom_marker_file = NULL,
  plot_umap = FALSE,
  name = "sctype_classification",
  assay = "SCT"
)
}
\arguments{
\item{seurat_object}{A Spatial Seurat object}

\item{known_tissue_type}{The tissue type of the input data should match what is in referenceDB}

\item{custom_marker_file}{Path to the custom marker file (optional)}

\item{plot_umap}{logical indicating whether to plot the UMAP (default= FALSE)}

\item{name}{The name of the metadata column to store the scType results (default is "sctype_classification")}

\item{assay}{e.g. RNA, SCT, integrated (default="SCT")}
}
\value{
A modified copy of the input Seurat object with a new metadata column
}
\description{
run an automated cell type annotation on spatial data-- spot type annotation
}
\details{
Useful to get an idea of different cells in the sample
}
\examples{
\dontrun{
seurat_object <- run_scType(seurat_object, known_tissue_type = "Hippo", assay = "SCT", name = "sctype-goes-spatial", custom_marker_file = "./ref_markers_brain.xlsx")
}
}
