% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_score.R
\name{sctype_score}
\alias{sctype_score}
\title{Calculate ScType scores and assign cell types}
\usage{
sctype_score(
  scRNAseqData,
  scaled = !0,
  gs,
  gs2 = NULL,
  gene_names_to_uppercase = !0,
  ...
)
}
\arguments{
\item{scRNAseqData}{input scRNA-seq matrix (rownames - genes, column names - cells),}

\item{scaled}{logical indicating whether the matrix is scaled (default=TRUE)}

\item{gs}{\itemize{
\item list of gene sets positively expressed in the cell type
}}

\item{gs2}{\itemize{
\item list of gene sets that should not be expressed in the cell type (NULL if not applicable) (default=NULL)
}}
}
\value{
scores
}
\description{
Calculate ScType scores and assign cell types
}
\author{
@IanevskiAleksandr
}
